requirements.txt
Flask==2.3.3
Jinja2==3.1.2
PyJWT==2.8.0
Werkzeug==2.3.7
python-dotenv==1.0.0
Pillow==10.0.1
pandas==2.1.1
openpyxl==3.1.2
selenium==4.15.2
